prototype/requirements.txt
ffmpeg-python>=0.2.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
tqdm>=4.66.0
